requirements.txt
protobuf==3.20.0
awscliv2==2.2.0
boto3==1.26.7
apache-airflow-providers-amazon==6.0.0
apache-airflow==2.4.3
pyspark==3.4.2
pytest-mock==3.11.1